[OPR-35]: Fixes assorted issues with Reachability
