+++ framework/Operations/Operations/DelayOperation.swift
@@ -50,7 +50,7 @@ public class DelayOperation: Operation {
 
         switch delay.interval {
 
-        case (let interval) where interval > 0:
+        case (let interval) where interval > 0.0:
             let after = dispatch_time(DISPATCH_TIME_NOW, Int64(interval * Double(NSEC_PER_SEC)))
             dispatch_after(after, Queue.Main.queue) {
                 if !self.cancelled {
